refactor: 주석 제거 및 파일명 변경
